Simplify selection of check for start and end_columns in ingest()
The code requires all four columns:
year_start, year_end, as well as date_start, date_end
Since it is possible to only provide two of them, the others need to be derived. The code previously was difficult to understand. This now simplifies the intent.

We also add a test in the unlikely case where all four are provided, but they are inconsistent.