Add NA placeholder for vpd in CRU daily downscaling
Since VPD is recomputed anyway on daily resolution it is not taken care of by `expand_clim_cru_monthly_byyr()`, i.e. it silently drops that column. More robust is if the column is kept and daily values are initialized as NA. This is done by this commit. Moreover, the data.frame() after recomputing VPD is now ungrouped, since this is less errror-prone.