Refrain from recomputing VPD for CRU based on potentially non-existing column 'vapr'
This fixes VPD for CRU, when vapr was not requested.
This commit moves computation of vpd with calc_vpd() inside of ingest_globalfields() for CRU, watch_wfdei, and wfde5. This results in consistent code between these 3 sources and also irrespective of the user calling ingest() or ingest_bysite(). Hence, this reduces code replication.
Furthermore, calc_vpd() is now re-computed (in ingest() or ingest_bysite()) only in cases when there is bias correction of `vapr` with WorldClim data. This code is still repeated across ingest() and ingest_bysite().